requirements.txt
bcrypt==4.2.1
blinker==1.9.0
boto3==1.36.6
botocore==1.36.6
click==8.1.8
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
packaging==24.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.2
six==1.17.0
urllib3
Werkzeug==3.1.3
